day_02/java/app/src/main/java/tkrippes/com/github/adventofcode2024/day02/parser/ReportListParser.java
package tkrippes.com.github.adventofcode2024.day02.parser;

import java.io.*;
import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

public class ReportListParser {
    public List<List<Integer>> parse(String inputFileName) throws IOException {
        File inputFile = new File(inputFileName);
        if (!inputFile.exists()) {
            throw new FileNotFoundException("Could not find input file: " + inputFileName);
        }

        List<List<Integer>> reportList = new ArrayList<>();

        try {
            BufferedReader reader = new BufferedReader(new FileReader(inputFile));
            String line;

            while ((line = reader.readLine()) != null) {
                reportList.add(Arrays.stream(line.split("\\s+")).map(Integer::parseInt).toList());
            }

            reader.close();
        } catch (IOException e) {
            throw new IOException("Error reading file: " + inputFile, e);
        }

        return reportList;
    }
}
